# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Shumi API",
        "title": "Shumi API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/namespace": {
            "put": {
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "Обновить информацию о неймспейсе",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название неймспейса",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Картинка неймспейса",
                        "name": "photo",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Описание неймспейса",
                        "name": "description",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Неймспейс успешно обновлен"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "Создать неймспейс",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название неймспейса",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Картинка неймспейса",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Описание неймспейса",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Неймспейс успешно создан"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Удалить неймспейс",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название неймспейса",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Неймспейс успешно удален"
                    },
                    "403": {
                        "description": "Этот пользователь не админ неймспейса",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            }
        },
...
        "/api/worker/autocomplete": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Предложенные пользователю интересы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ввод пользователя",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.AutocompleteResponse"
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "database.Error": {
            "description": "Ошибка",
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.AutocompleteResponse": {
            "description": "Ответ autocomplete эндпоинта воркера.",
            "type": "object",
            "properties": {
                "q": {
                    "type": "string"
                },
                "response": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}